treestructure: Detect Population Structure Within Phylogenetic Trees

Algorithms for detecting population structure from the history of coalescent events recorded in phylogenetic trees. This method classifies each tip and internal node of a tree into disjoint sets characterized by similar coalescent patterns.

Details

Methods for detecting structure in phylogenies. Includes the *trestruct* function for partitioning a tree and methods for printing and plotting trees with structure. Refer to the vignettes for common usage.

References

Erik Volz, Wiuf Carsten, Yonatan Grad, Simon Frost, Ann Dennis, Xavier Didelot, "Identification of hidden population structure in time-scaled phylogenies", (2020); Systematic Biology, 69: 884–896.

See also

Author

Maintainer: Erik Volz erik@erikvolz.info (ORCID)

Other contributors:

Fabricia F. Nascimento thednainus@yahoo.com (ORCID) [contributor]

Vinicius B. Franceschi vinibfranc@gmail.com (ORCID) [contributor]